Only delete .java and .class files when cleaning the gen folder (#336) also changed the default for the destination directory to ./gen/
git-svn-id: https://repos.hats-project.eu:444/svn/hats/Tools/ABS/trunk@13738 97f52227-7fa9-4375-a72d-c1c733b1f48f

+++ frontend/src/abs/backend/java/codegeneration/JavaGeneratorHelper.java
@@ -344,7 +344,10 @@ public class JavaGeneratorHelper {
             if (f.isDirectory()) {
                 cleanGenFolderRecursively(f);
             } else {
-                f.delete();
+                if (f.getAbsolutePath().endsWith(".java")
+                        || f.getAbsolutePath().endsWith(".class")) {
+                    f.delete();
+                }
             }
         }
         dir.delete();
